bugz: separate bugzilla datatypes into common package.
This helps decoupling bugz package from database package and yet sharing
data types.
